+++ src/components/index.ts
@@ -1,2 +1,6 @@
 export * from './Email';
+export * from './Section';
+export * from './Column';
+export * from './Button';
+export * from './Link';
 export * from './Divider';